Add files via upload
File takes 90 minutes to run in full from 2009.  Takes a more reasonable amount of time from around 2016.  The ticker list are the 40 Current or removed DJIA constituents since 2009.  Intel and GE seem to have filed their 10-Q in a format the has incorrect spacing and it encoded as only having 1 item instead of several items.  I can't figure out how to parse through it but it is not material as long as we're okay with missing their text data since around 2015.